Extract src-tauri/src/platform.rs — the OS-shaped leaves
Phase 2's first slice, and a reorder: PLAN listed platform.rs last, but the
greps say it must come first. sys_command has 18 uses, norm_path 12, sh_quote
11, home_dir 9, and git_cmd itself calls augmented_path — so every other
Phase-2 module depends on this one. That is Phase 1's seam rule 1 (move the
callee down first), the same reason icons.ts preceded sidebar.ts.

Moved: home_dir, norm_path, sys_command, resolve_claude (both cfg arms),
augmented_path (both), sh_quote, ps_one (both), plus the three tests over
norm_path and sh_quote. 173 lines. lib.rs 4,711 -> 4,532.

Deliberately NOT moved, and the rule behind it: a cfg-gated helper with a
single consumer module belongs to that module. apply_utf8_locale,
interactive_shell and task_shell are held for pty.rs — apply_utf8_locale takes
a portable_pty::CommandBuilder, which is the tell that the leaf layer would
have to import portable_pty to hold it. same_path is held for git.rs.

Also records the AppState/Session decision the kickoff prompt asked for: they
stay in the crate root as crate::AppState, no state.rs. run() is their only
constructor and lives in lib.rs; pub(crate) plus the compiler does what Phase 1
needed seven setX functions for; and two structs is a file per 34 lines.

Verification, both halves, line-sorted:
- the 173 moved lines vs `git show HEAD:src-tauri/src/lib.rs` — identical
  modulo the pub(crate) prefixes, with one intended difference: ps_one's doc
  comment said "the cross-platform ProcTable below", which the move made false.
- the remainder of lib.rs vs HEAD-minus-those-ranges — the only differences are
  the four wiring lines added by hand (mod platform; and the two use lines).

cargo test: 69 passed, same as baseline. cargo check --tests clean, no
warnings. No TypeScript touched.
